+++ Editor/Drawers/RawReferenceDrawer.cs
@@ -8,8 +8,11 @@ namespace TNRD.Drawers
 {
     internal class RawReferenceDrawer : ReferenceDrawer, IReferenceDrawer
     {
-        private readonly GUIContent label;
-        private readonly FieldInfo fieldInfo;
+        private GUIContent label;
+        private FieldInfo fieldInfo;
+
+        private object previousReferenceValue;
+        private string previousPropertyPath;
 
         private object RawReferenceValue
         {
@@ -23,12 +26,20 @@ namespace TNRD.Drawers
                 return instance.GetRawReference();
 #endif
             }
+
+            set
+            {
+#if UNITY_2021_1_OR_NEWER
+                RawReferenceProperty.managedReferenceValue = value;
+#else
+                fieldInfo.SetValue(Property.serializedObject.targetObject, value);
+#endif
+            }
         }
 
-        /// <inheritdoc />
-        public RawReferenceDrawer(SerializedProperty property, GUIContent label, Type genericType, FieldInfo fieldInfo)
-            : base(property, genericType)
+        public void Initialize(SerializedProperty property, Type genericType, GUIContent label, FieldInfo fieldInfo)
         {
+            Initialize(property, genericType);
             this.label = label;
             this.fieldInfo = fieldInfo;
         }
@@ -47,6 +58,8 @@ namespace TNRD.Drawers
         /// <inheritdoc />
         public void OnGUI(Rect position)
         {
+            AvoidDuplicateReferencesInArray();
+
             Rect objectFieldRect = new Rect(position)
             {
                 height = EditorGUIUtility.singleLineHeight
@@ -72,6 +85,8 @@ namespace TNRD.Drawers
                 RawReferenceProperty,
                 new GUIContent(rawReferenceValue.GetType().Name),
                 true);
+
+            previousPropertyPath = Property.propertyPath;
         }
 
         /// <inheritdoc />
@@ -91,6 +106,42 @@ namespace TNRD.Drawers
                     return;
                 }
             }
+        }
+
+        private void AvoidDuplicateReferencesInArray()
+        {
+            if (!IsPropertyInArray(Property)) 
+                return;
+            if (previousPropertyPath == null) 
+                return;
+            if (previousPropertyPath == Property.propertyPath) 
+                return;
+
+            SerializedProperty rawReferenceProperty = Property.FindPropertyRelative("rawReference");
+            object currentReferenceValue = RawReferenceValue;
+
+            if (currentReferenceValue == null) 
+                return;
+
+            if (previousReferenceValue == currentReferenceValue)
+            {
+                RawReferenceValue = CreateInstance(currentReferenceValue);
+                rawReferenceProperty.serializedObject.ApplyModifiedProperties();
+            }
+
+            previousReferenceValue = currentReferenceValue;
+        }
+
+        private static bool IsPropertyInArray(SerializedProperty prop)
+        {
+            return prop.propertyPath.Contains(".Array.data[");
+        }
+
+        private static object CreateInstance(object source)
+        {
+            object instance = Activator.CreateInstance(source.GetType());
+            EditorUtility.CopySerializedManagedFieldsOnly(source, instance);
+            return instance;
         }
     }
 }